k8s: extract metrics related code in separate metrics.go file
To reduce the size and scope of the k8s/watchers/watcher.go file,
this commit extracts all metrics relevant structs and their
initialization/registration into a separate file metrics.go.
